Shopping-Agent/requirements.txt
langchain-core>=0.3.60
langchain-openai>=0.3.21
langchain-anthropic>=0.3.0
langchain-google-genai>=2.0.5
langchain-community>=0.3.0
python-dotenv>=0.21.0
requests>=2.32.0